Modify encounter server functions to serve the username with the encounters
Before the encounters were just getting served up with the userid, but that wasn't very descriptive. I changed the helper functions to get requests to encounter endpoints returned the encounters and the name of the user the encounters were associated with. I had to alter the front-end to handle the change in data structure.